fix: move env override test from tests/core to tests/config
The test imports autoskillit.config._config_loader which violates
the IL-0 layer boundary enforced by test_layer_enforcement.
