+++ tests/config/test_config.py
@@ -583,6 +583,18 @@ def test_secrets_only_keys_covers_all_github_secret_fields() -> None:
             )
 
 
+def test_experimental_enabled_env_override_false(monkeypatch: pytest.MonkeyPatch) -> None:
+    """AUTOSKILLIT_FEATURES__EXPERIMENTAL_ENABLED=false disables experimental features."""
+    from autoskillit.core.feature_flags import is_feature_enabled
+
+    monkeypatch.setenv("AUTOSKILLIT_FEATURES__EXPERIMENTAL_ENABLED", "false")
+    cfg = load_config()
+    assert not cfg.experimental_enabled
+    assert not is_feature_enabled(
+        "fleet", cfg.features, experimental_enabled=cfg.experimental_enabled
+    )
+
+
 class TestWorkspaceConfig:
     """WorkspaceConfig section is present in AutomationConfig with correct defaults."""
